docker_commander/requirements.txt
python-telegram-bot>=22.0
httpx>=0.27.0
PyYAML>=6.0